+++ payment-api/src/main/java/org/ecommerce/paymentapi/external/controller/BeanPayController.java
@@ -10,7 +10,6 @@ import org.ecommerce.paymentapi.dto.request.TossFailRequest;
 import org.ecommerce.paymentapi.dto.request.TossPaymentRequest;
 import org.ecommerce.paymentapi.dto.response.PaymentDetailResponse;
 import org.ecommerce.paymentapi.external.service.BeanPayService;
-import org.ecommerce.paymentapi.external.service.PaymentDetailReadService;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -29,7 +28,6 @@ import lombok.extern.slf4j.Slf4j;
 public class BeanPayController {
 
 	private final BeanPayService beanPayService;
-	private final PaymentDetailReadService paymentDetailReadService;
 
 	@PostMapping("/charge")
 	public Response<PaymentDetailResponse> beforeCharge(@RequestBody final PreChargeRequest request) {
@@ -62,57 +60,5 @@ public class BeanPayController {
 			PaymentDetailMapper.INSTANCE.toResponse(response)
 		);
 	}
-
-	@GetMapping("/user")
-	public Response<Page<PaymentDetailResponse>> getUserPayments(
-		@RequestParam String startDateTime,
-		@RequestParam String endDateTime,
-		@RequestParam(required = false) PaymentStatus status,
-		Pageable pageable
-	) {
-		LocalDateTime start = PaymentTimeFormatUtil.stringToDateTime(startDateTime);
-		LocalDateTime end = PaymentTimeFormatUtil.stringToDateTime(endDateTime);
-		//TODO: jwt userId 주입 예정
-		Page<PaymentDetailDto> paymentDetailDtoPage =
-			paymentService.getUserPaymentDetailsByBetweenDate(
-				999, start, end, status, pageable
-			);
-		return new Response<>(
-			HttpStatus.OK.value(),
-			new PageImpl<>(
-				paymentDetailDtoPage.getContent().stream()
-					.map(PaymentMapper.INSTANCE::toPaymentDetailResponse)
-					.toList(),
-				pageable,
-				paymentDetailDtoPage.getTotalPages())
-		);
-	}
-
-	@GetMapping("/seller")
-	public Response<Page<PaymentDetailResponse>> getSellerPayments(
-		@RequestParam String startDateTime,
-		@RequestParam String endDateTime,
-		@RequestParam(required = false) PaymentStatus status,
-		Pageable pageable
-	) {
-		LocalDateTime start = PaymentTimeFormatUtil.stringToDateTime(startDateTime);
-		LocalDateTime end = PaymentTimeFormatUtil.stringToDateTime(endDateTime);
-		//TODO: jwt sellerId 주입 예정
-		Page<PaymentDetailDto> paymentDetailDtoPage =
-			paymentService.getSellerPaymentDetailByBetweenRange(
-				1000, start, end, status, pageable
-			);
-		return new Response<>(
-			HttpStatus.OK.value(),
-			new PageImpl<>(
-				paymentDetailDtoPage.getContent().stream()
-					.map(PaymentMapper.INSTANCE::toPaymentDetailResponse)
-					.toList(),
-				pageable,
-				paymentDetailDtoPage.getTotalPages())
-		);
-	}
-
-
 
 }
